FCW/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
django-heroku==0.3.1
django-jet-reboot==1.1.1
django-paypal==1.1.1
djangorestframework==3.12.4
docx==0.2.4
feedparser==6.0.8
gunicorn==20.1.0
idna==2.10
lxml==4.6.3
Pillow==7.2.0
psycopg2==2.9.1
PyMySQL==1.0.2
python-docx==0.8.10
pytz==2020.1
reportlab==3.5.67
requests==2.25.1
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.3.1
typing-extensions==3.10.0.0
urllib3==1.26.5
whitenoise==5.2.0
asgiref==3.7.2
Django==4.2.7
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
